requirements.txt
gradio==3.50.2
tensorflow-cpu>=2.15.0,<2.16.0
keras>=2.15.0,<2.16.0
pillow>=10.0.0
numpy>=1.24.0
